Убрал Service-слой, создал ActionListener отдельно для каждой кнопки, заприватил некоторые поля, отделил реализацию кнопок в PaintWindow.
